# api/openapi.yaml
openapi: 3.1.0
info:
  title: skali API
  description: |
    Client-facing API of skali. Authentication is email+password with
    optional TOTP two-factor; sessions are opaque bearer tokens. The same
    API serves every client: the web BFF, the `skali` CLI, and future
    native apps.

    There is no registration endpoint: users are created by an instance
    admin (`/v1/users`) or by the server operator (`skalid user create`).
    Every user carries an instance-wide role, `admin` (everything,
    including users, nodes, and instance settings) or `member` (nothing
    until granted). Product access is per project and per environment on
    one ladder of roles, `none < read < deploy < maintain < admin`: a
    project membership is the member's default role on every environment
    of the project, a per-environment cell overrides it for one user, and
    an environment ceiling (`max_role`) caps what members inherit.
    Non-members do not see a project at all (`404`); a locked environment
    (`none`) is listed by name but its contents answer `403 forbidden`
    naming the required role. Project and environment payloads carry the
    caller's `access` so clients can disable and explain without guessing.
    All error responses share the `Error` envelope; clients should branch
    on `error.code`, not on messages.

    Security-critical operations sit behind sudo mode: they require the
    session to have proven the user's identity recently (within the
...

...

...
  version: 0.1.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/projects:
    post:
      operationId: createProject
      summary: Create a project
      description: |
        Instance admins always; members need `create_projects`. The
        creator becomes the project's admin.
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  pattern: "^[a-z][a-z0-9-]{0,62}$"
                display_name:
                  type: string
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
    get:
      operationId: listProjects
      summary: List projects
      tags: [projects]
      parameters:
        - name: include
          in: query
          required: false
          schema:
            type: string
            enum: [summary]
          description: |
            With `summary`, each project carries its environments (state
            plus worst service health) and draft service counts.
      responses:
        "200":
          description: |
            The caller's projects (memberships; instance admins see all),
            sorted by name.
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items: { $ref: "#/components/schemas/Project" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getProject
      summary: Get a project
      tags: [projects]
      responses:
        "200":
          description: The project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateProject
      summary: Update a project
      description: |
        Updates the display name and/or the draft source mode. Switching
        the source mode is an explicit action with product consequences:
        `managed` documents are edited through the web/API, `file`
        documents are replaced by CLI/Git submissions.
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                source_mode:
                  type: string
                  enum: [managed, file]
      responses:
        "200":
          description: Updated project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteProject
      summary: Delete a project
      description: |
        Destructive: removes the project with every environment, draft,
        and revision. Requires recent authentication (sudo mode).
      tags: [projects]
      responses:
        "204":
          description: Project deleted.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/ReauthRequired" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/projects/{id}/definitions:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: submitDefinition
      summary: Submit a candidate definition version
      description: |
        Compiles and stores a definition version without moving the project
        draft. Deployment preparation consumes the returned id; only
        promotion advances the draft. Identical content reuses the stored
        version.
      tags: [deployments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source:
                  type: string
                format:
                  type: string
                  enum: [yaml, json]
      responses:
        "200":
          description: The stored definition version.
          content:
            application/json:
              schema:
                type: object
                required: [definition_version_id, definition_hash]
                properties:
                  definition_version_id:
                    type: string
                    format: uuid
                  definition_hash:
                    type: string
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: The manifest failed validation (`invalid_manifest`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token from login / 2FA verify.
  responses:
    BadRequest:
      description: Malformed request body or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing/invalid token, wrong credentials, or wrong code.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        The caller is authenticated but lacks the role the operation needs
        (`error.code = forbidden`; the message names it, for example
        "maintain on environment production required"), or, on
        sudo-mode-gated operations, recent authentication is required
        (`error.code = reauth_required`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        The resource does not exist, or the caller is not a member of the
        project it belongs to (the two are indistinguishable by design).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: The resource is in a state that does not allow the operation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ReauthRequired:
      description: |
        Recent authentication required (`error.code = reauth_required`):
        the session's sudo-mode window has lapsed. Confirm identity at
        `/v1/auth/reauth` and retry the original request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    AccessRole:
      type: string
      description: |
        One step of the access ladder, each including everything below it.
        `none` locks an environment (listed by name, contents refused);
        `read` sees everything; `deploy` changes only what code runs
        (promote, rollback, restart, cancel, backup create, direct deploys
        whose definition is unchanged); `maintain` owns the blast radius of
        skali.yml (definition changes, values, restore, exec, credential
        reveal); `admin` manages settings, other users' roles, delete and
        teardown. Project memberships use `read` and up.
      enum: [none, read, deploy, maintain, admin]
    ProjectAccess:
      type: object
      description: |
        The caller's standing on a project: the membership role (instance
        admins report `admin`) and the effective role per environment name
        (locked environments report `none`).
      required: [role, environments]
      properties:
        role: { $ref: "#/components/schemas/AccessRole" }
        environments:
          type: object
          additionalProperties: { $ref: "#/components/schemas/AccessRole" }
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - invalid_credentials
            - invalid_token
            - invalid_code
            - forbidden
            - reauth_required
            - not_found
            - conflict
            - version_conflict
            - invalid_manifest
            - rate_limited
            - node_unreachable
            - no_ready_pod
            - registry_disabled
            - registry_unavailable
            - internal
        message:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/ErrorDetail" }
    Project:
      type: object
      required: [id, name, display_name, source_mode, created_at, updated_at, access]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Stable identity; must match the manifest `name`.
        display_name:
          type: string
        source_mode:
          type: string
          enum: [managed, file]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        access: { $ref: "#/components/schemas/ProjectAccess" }
        summary:
          $ref: "#/components/schemas/ProjectSummary"
    ProjectSummary:
      type: object
      description: |
        List rollup attached by `?include=summary`: environments with their
        pointer state and worst service health, plus the draft's per-type
        service counts (zeros without a draft). A locked environment
        (`access: none`) carries id, name, and access only.
      required: [environments, service_counts]
      properties:
        environments:
          type: array
          items:
            type: object
            required: [id, name, access]
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              access: { $ref: "#/components/schemas/AccessRole" }
              state:
                type: string
                enum: [active, down, releasing]
              health:
                type: string
                enum: [unknown, progressing, healthy, degraded, unhealthy]
        service_counts:
          type: object
          required: [applications, databases, buckets]
          properties:
            applications:
              type: integer
            databases:
              type: integer
            buckets:
              type: integer
    ProjectResponse:
      type: object
      required: [project]
      properties:
        project: { $ref: "#/components/schemas/Project" }
...